API Reference

Calling functions and procedures

MagmaCall.magcall — Function

magcall(Val(C), f, args...; opts...)

Call the Magma function f with the given arguments and keyword options according to calling convention C.

If f is a Symbol, then it is interpreted as the name of a Magma intrinsic.

C may be:

:p: f is a procedure and nothing is returned. Note that if f is actually a function, its return value will be printed.

:f: f is a function and its first return value is returned.

:fN: f is a function and its first N return values are returned as a tuple.

:b: f is a function returning a boolean, which is returned as a Bool.

:i: f is a function returning an integer, which is returned as an Int.

:g: f is a function returning a structure S. args[1] is a tuple of generator names, which are applied to S. Returns a tuple of S and generator names.

:pN: Same as :p (procedure) but the Nth argument is passed by reference.

Equivalently, you can call magcallC(f, args...; opts...) or magC.f(args...; opts...).

MagmaCall.magcallp — Function

magcallp(f, args...; opts...)

Call the Magma procedure f with the given arguments and keyword options.

Equivalent to magcall(Val(:p), f, args...; opts...).

MagmaCall.magcallf — Function

magcallf(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Return the first return value.

Equivalent to magcall(Val(:f), f, args...; opts...).

MagmaCall.magcallf2 — Function

magcallf2(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Return the first two return values.

Equivalent to magcall(Val(:f2), f, args...; opts...).

MagmaCall.magcallf3 — Function

magcallf3(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Return the first three return values.

Equivalent to magcall(Val(:f3), f, args...; opts...).

MagmaCall.magcallf4 — Function

magcallf4(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Return the first four return values.

Equivalent to magcall(Val(:f4), f, args...; opts...).

MagmaCall.magcallf5 — Function

magcallf5(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Return the first five return values.

Equivalent to magcall(Val(:f5), f, args...; opts...).

MagmaCall.magcallf6 — Function

magcallf6(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Return the first six return values.

Equivalent to magcall(Val(:f6), f, args...; opts...).

MagmaCall.magcallf7 — Function

magcallf7(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Return the first seven return values.

Equivalent to magcall(Val(:f7), f, args...; opts...).

MagmaCall.magcallf8 — Function

magcallf8(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Return the first eight return values.

Equivalent to magcall(Val(:f8), f, args...; opts...).

MagmaCall.magcalli — Function

magcalli(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. It must return an integer, which is returned as an Int.

Equivalent to magcall(Val(:i), f, args...; opts...).

MagmaCall.magcallb — Function

magcallb(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. It must return a boolean, which is returned as a Bool.

Equivalent to magcall(Val(:b), f, args...; opts...).

MagmaCall.magcallp1 — Function

magcallp1(f, args...; opts...)

Call the Magma procedure f with the given arguments and keyword options, passing the first argument by reference.

Equivalent to magcallp1(Val(:p1), args...; opts...).

MagmaCall.magcallp2 — Function

magcallp2(f, args...; opts...)

Call the Magma procedure f with the given arguments and keyword options, passing the second argument by reference.

Equivalent to magcallp2(Val(:p2), args...; opts...).

MagmaCall.magcallp3 — Function

magcallp3(f, args...; opts...)

Call the Magma procedure f with the given arguments and keyword options, passing the third argument by reference.

Equivalent to magcallp3(Val(:p3), args...; opts...).

MagmaCall.magcallp4 — Function

magcallp4(f, args...; opts...)

Call the Magma procedure f with the given arguments and keyword options, passing the fourth argument by reference.

Equivalent to magcallp4(Val(:p4), args...; opts...).

MagmaCall.magcallg — Function

magcallg(f, names, args...; opts...)

Call the Magma function f with the given arguments and keyword options. Assign the given names to the result. Return a tuple of the result and the generators.

Equivalent to magcall(Val(:g), f, names, args...; opts...).

MagmaCall.magcallm — Function

magcallm(f, args...; opts...)

Call the Magma function f with the given arguments and keyword options. If the first return value is true, return the second return value, else return nothing.

Equivalent to magcall(Val(:m), args...; opts...).

Binary operators

MagmaCall.magadd — Function

magadd(x, y)

Equivalent to x + y in Magma.

MagmaCall.magsub — Function

magsub(x, y)

Equivalent to x - y in Magma.

MagmaCall.magmul — Function

magmul(x, y)

Equivalent to x * y in Magma.

MagmaCall.magtruediv — Function

magtruediv(x, y)

Equivalent to x / y in Magma.

MagmaCall.magpow — Function

magpow(x, y)

Equivalent to x ^ y in Magma.

MagmaCall.magdiff — Function

magdiff(x, y)

Equivalent to x diff y in Magma.

MagmaCall.magdiv — Function

magdiv(x, y)

Equivalent to x div y in Magma.

MagmaCall.magjoin — Function

magjoin(x, y)

Equivalent to x join y in Magma.

MagmaCall.magmeet — Function

magmeet(x, y)

Equivalent to x meet y in Magma.

MagmaCall.magmod — Function

magmod(x, y)

Equivalent to x mod y in Magma.

MagmaCall.magsdiff — Function

magsdiff(x, y)

Equivalent to x sdiff y in Magma.

MagmaCall.magcat — Function

magcat(x, y)

Equivalent to x cat y in Magma.

MagmaCall.mageq — Function

mageq(x, y)

True if x eq y in Magma.

MagmaCall.magne — Function

magne(x, y)

True if x ne y in Magma.

MagmaCall.magcmpeq — Function

magcmpeq(x, y)

True if x cmpeq y in Magma.

MagmaCall.magcmpne — Function

magcmpne(x, y)

True if x cmpne y in Magma.

MagmaCall.magle — Function

magle(x, y)

True if x le y in Magma.

MagmaCall.maglt — Function

maglt(x, y)

True if x lt y in Magma.

MagmaCall.magge — Function

magge(x, y)

True if x ge y in Magma.

MagmaCall.maggt — Function

maggt(x, y)

True if x gt y in Magma.

MagmaCall.magin — Function

magin(x, y)

True if x in y in Magma.

MagmaCall.magnotin — Function

magnotin(x, y)

True if x notin y in Magma.

MagmaCall.magadj — Function

magadj(x, y)

True if x adj y in Magma.

MagmaCall.magnotadj — Function

magnotadj(x, y)

True if x notadj y in Magma.

MagmaCall.magsubset — Function

magsubset(x, y)

True if x subset y in Magma.

MagmaCall.magnotsubset — Function

magnotsubset(x, y)

True if x notsubset y in Magma.

Reductions

MagmaCall.magsum — Function

magsum(x)

Equivalent to &+ x in Magma.

MagmaCall.magprod — Function

magprod(x)

Equivalent to &* x in Magma.

MagmaCall.magreduceand — Function

magreduceand(x)

Equivalent to &and x in Magma.

MagmaCall.magreduceor — Function

magreduceor(x)

Equivalent to &or x in Magma.

MagmaCall.magreducejoin — Function

magreducejoin(x)

Equivalent to &join x in Magma.

MagmaCall.magreducemeet — Function

magreducemeet(x)

Equivalent to &meet x in Magma.

MagmaCall.magreducecat — Function

magreducecat(x)

Equivalent to &cat x in Magma.

Containers

MagmaCall.maglist — Function

maglist([xs])

A new Magma list, equivalent to [* *].

Optionally, values can be specified with iterable xs.

MagmaCall.magtuple — Function

magtuple([xs])

A new Magma tuple, equivalent to < >.

Optionally, values can be specified with iterable xs.

MagmaCall.magseq — Function

magseq([xs]; [universe])

A new Magma sequence, equivalent to [ universe | ].

Optionally, values can be specified with iterable xs.

MagmaCall.magset — Function

magset([xs]; [universe])

A new Magma set, equivalent to { universe | }.

Optionally, values can be specified with iterable xs.

MagmaCall.magmset — Function

magmset([xs]; [universe])

A new Magma multiset, equivalent to {* universe | *}.

Optionally, values can be specified with iterable xs.

MagmaCall.magiset — Function

magiset([xs]; [universe])

A new Magma indexed set, equivalent to {@ universe | @}.

Optionally, values can be specified with iterable xs.

Constructors

MagmaCall.magrecformat — Function

magrecformat(field, ...)

Create a record format with the given fields.

Each field may be a symbol, giving the name, or a name=>type pair.

MagmaCall.magrec — Function

magrec(fmt; vals...)

A record with the given format and values.

Other operators

MagmaCall.magprint — Function

magprint(io::IO, x, level=:default)

Print x to io at the given level, which must be one of :default, :magma, :maximal or :minimal.

MagmaCall.magsprint — Function

magsprint(x, level=:default)

String representation of x at the given level (see magprint).

MagmaCall.maggetindex — Function

maggetindex(o, i, ...)

Equivalent to o[i, ...] in Magma.

MagmaCall.magsetindex! — Function

magsetindex!(o, x, i...)

Equivalent to o[i, ...] := x in Magma.

MagmaCall.maglength — Function

maglength(o)

The length of o, equivalent to # o.

MagmaCall.magimage — Function

magimage(x, f)

The image of x under f, equivalent to x @ f.

MagmaCall.magpreimage — Function

magpreimage(x, f)

The preimage of x under f, equivalent to x @@ f.

MagmaCall.magcoerce — Function

magcoerce(s, x)

Coerce x into structure s, equivalent to s ! x.

MagmaCall.maggen — Function

maggen(x, n)

The nth generator of x, equivalent to x.n.

MagmaCall.maggetattr — Function

maggetattr(x, k)

Get attribute k from x, equivalent to x`k.

MagmaCall.magsetattr! — Function

magsetattr!(x, k, v)

Set attribute k of x to v, equivalent to x`k := v.

MagmaCall.magattrnames — Function

magattrnames(x)

The attribute names of x.

MagmaCall.mageval — Function

mageval(ex; vars...)

Evaluate the Magma expression ex and return its value.

Keyword arguments can be used to name variables refereced by ex, e.g. mageval("x+x", x=3) returns 6.

This is the worker function behind @mag_str.

MagmaCall.maghelp — Function

maghelp(query)

Equivalent to ? query.

MagmaCall.magassigned — Function

magassigned(o::MagmaObject)

True if o is assigned. This should only occur if o = MagmaObject() or if magdelete(o) has been called.

Most functions do not expect unassigned values, and undefined behaviour occurs.

MagmaCall.magdelete — Function

magdelete(o::MagmaObject)

Delete the object referenced by o.

Most functions do not expect deleted values, and undefined behaviour occurs.

Other helpers

MagmaCall.magtypeof — Function

magtypeof(o)

The Magma type of o.

MagmaCall.magetypeof — Function

magetypeof(o)

The Magma extended type of o.

MagmaCall.magissubtype — Function

magissubtype(t1, t2)

True if Magma type t1 is a subtype of t2.

Symbol arguments are interpreted as types.

MagmaCall.magisinstance — Function

magisinstance(o, t)

True if o is an instance of Magma type t.

If t is a Symbol, it is interpreted as a type.

Magma values and conversion

MagmaCall.MagmaValue — Type

MagmaValue <: Any
MagmaValue(x)

Abtract type for any kind of Magma value. Most functions accept any MagmaValue and return a MagmaObject.

Subtypes include MagmaObject, MagmaRef, MagmaExpr.

Overload MagmaValue(x) to provide conversions from Julia values to Magma ones.

MagmaCall.MagmaObject — Type

MagmaObject <: MagmaValue
MagmaObject(x)

A Magma object value. Most functions return a MagmaObject.

It is represented by a handle to a variable in the Magma interpreter.

MagmaCall.MagmaRef — Type

MagmaRef <: MagmaValue
MagmaRef(o::MagmaObject)

A reference to a Magma object, suitable for passing to mutating procedures such as Append or Sort.

MagmaCall.MagmaExpr — Type

MagmaExpr <: MagmaValue
MagmaExpr(ex::String)

A literal Magma value, represented by an expression which evaluates to the value.

For simple values such as integers and strings, MagmaValue(x) will return a MagmaExpr.

MagmaCall.MagmaIntrinsic — Type

MagmaIntrinsic <: MagmaValue
MagmaIntrinsic(name)

The Magma intrinsic with the given name.

MagmaCall.MagmaType — Type

MagmaType <: MagmaValue
MagmaType(name, ...)

The Magma type with the given name.

Extra arguments make an extended type, e.g. MagmaType(:RngUPol, :RngInt) is Magma's RngUPol[RngInt].

MagmaCall.MagmaCallable — Type

MagmaCallable{C} <: MagmaValue
MagmaCallable{C}(o)

Wrap the value o, declaring it to be callable with calling convention C (see magcall).

The resulting object can use the usual function call syntax, instead of magcall.

MagmaCall.magconvert — Function

magconvert(T, o)

Convert the Magma object o to a T.

MagmaCall.@mag_str — Macro

mag"..."[C]

Evaluates the given string as a Magma expression.

Julia values may be interpolated with $ as usual.

If C is given, the returned value is wrapped as a MagmaCallable{C}, for example:

f = mag"func<x, y | x^2 + 3*x*y + y^2>"f
f(2, 3)